After FIVE DAYS of no updates, constant failure and several meltdowns, I have finally done something that should have been the easiest thing on planet Earth: making drawing possible in VR. :]
Now, on to seeing if the if I could implement the q-dollar gesture recognition for the glyphs!

Other than that, I've been playing around with in-game UIs, as you will be able to see once you launch the game. The UI elements won't disappear with the book, though, I already have a fix for that (methinks), but I haven't implemented it yet.
And before you say, no I can't just tie it to the book as the Node's "box" messes with the Function Pointers for some reason, idk.

Side note here: The Scenes are a MESS, I am in the middle of moving everything from regular Raycasts to FunctionPointers as they were literally DESIGNED for VR and I just didn't know about it, because the extention has really bad documentation, still props to the guys who made all this, so I wouldn't have to, thanks :D